refactor: split activate endpoint integration tests from endpoints suite
Move activate API coverage into a dedicated integration file and share HTTP test helpers so endpoint suites stay focused and avoid repeated max-lines pressure.
